Add fromstdinv2
I have a new version of the argparse code to pull in values from
a pipe. This sub classes the built in store action. This is more
code then the first version but would be easier to import and
implement.

I am also renaming the fromstdin.py file to fromstdinv1 so that
I can better denote the different versions.